pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aioconsole>=0.8.1",
    "aiohttp>=3.12.11",
    "elasticsearch>=9.0.1",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "mcp[cli]>=1.9.2",
    "openai>=1.82.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
